Separate PixelSnapper from Button.cc
