+++ src/ShadeButton.h
@@ -43,35 +43,32 @@ public:
     static void paintIcon(Button *button, QPainter *painter, const QRectF &iconRect, const qreal dpr) {
         Q_UNUSED(iconRect)
         PixelSnapper snapper(painter, dpr);
-        auto snapPoint = [&snapper](const QPointF &p) -> QPointF {
-            return snapper.snap(p);
-        };
 
         const QPointF offset(-5.0, -5.0);
 
         if (button->isChecked()) {
             button->setPenWidth(painter, 1.25);
             painter->drawLine( 
-                snapPoint(QPointF( 0, 2 ) + offset),
-                snapPoint(QPointF( 10, 2 ) + offset)
+                snapper.snap(QPointF( 0, 2 ) + offset),
+                snapper.snap(QPointF( 10, 2 ) + offset)
             );
             button->setPenWidth(painter, 1.25);
             painter->drawPolyline(  QVector<QPointF> {
-                snapPoint(QPointF( 0.0, 5.0 ) + offset),
-                snapPoint(QPointF( 5.0, 10.0 ) + offset),
-                snapPoint(QPointF( 10.0, 5.0 ) + offset)
+                snapper.snap(QPointF( 0.0, 5.0 ) + offset),
+                snapper.snap(QPointF( 5.0, 10.0 ) + offset),
+                snapper.snap(QPointF( 10.0, 5.0 ) + offset)
             });
         } else {
             button->setPenWidth(painter, 1.25);
             painter->drawLine( 
-                snapPoint(QPointF( 0, 2 ) + offset),
-                snapPoint(QPointF( 10, 2 ) + offset)
+                snapper.snap(QPointF( 0, 2 ) + offset),
+                snapper.snap(QPointF( 10, 2 ) + offset)
             );
             button->setPenWidth(painter, 1.25);
             painter->drawPolyline( QVector<QPointF> {
-                snapPoint(QPointF( 0.0, 10.0 ) + offset),
-                snapPoint(QPointF( 5.0, 5.0 ) + offset),
-                snapPoint(QPointF( 10.0, 10.0 ) + offset)
+                snapper.snap(QPointF( 0.0, 10.0 ) + offset),
+                snapper.snap(QPointF( 5.0, 5.0 ) + offset),
+                snapper.snap(QPointF( 10.0, 10.0 ) + offset)
             });
         }
     }
